requirements.txt
# Lightweight, CPU-only stack
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers>=4.40
sentence-transformers
faiss-cpu
PyMuPDF            # fast PDF parser
tqdm